requirements.txt
africastalking==1.2.9
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
coverage==7.6.10
cryptography==44.0.0
Django==5.1.5
django-allauth==65.3.1
django-graphql-jwt==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
josepy==1.15.0
promise==2.3
psycopg==3.2.4
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
schema==0.7.7
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.3.0